- EXACTIP-branch: merged all changes from trunk (rootof-EXACTIP -- v1108)

+++ updatedates.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 for i in Makefile INSTALL check/*.sh check/*.awk doc/*.c src/*.c src/*.cpp src/*.h src/blockmemshell/*.c src/blockmemshell/*.h src/objscip/*.cpp src/objscip/*.h src/scip/*.c src/scip/*.cpp src/scip/*.h src/tclique/*.c src/tclique/*.h
 do
+if [ -f $i ]
+then
 echo $i
 mv $i $i.olddate
-sed 's!2002-2006!2002-2007!g
-s!1996-2006!1996-2007!g' $i.olddate > $i
+sed 's!2002-2008!2002-2009!g
+s!1996-2008!1996-2009!g' $i.olddate > $i
+fi
 done
